feat(scouts): shared ScoutRowCard on fleet list and detail screen
Extract the fleet row into ScoutRowCard and reuse it as the header
card on the scout detail screen (linkToDetail=false there), replacing
the old title block + Configuration card so the two surfaces stay
identical. Also drop the quiet-run sentence from the Scouts
subsection description.
