+++ packages/ui/src/features/scouts/components/ScoutsFleetSection.tsx
@@ -1,17 +1,8 @@
-import {
-  ArrowSquareOutIcon,
-  CaretDownIcon,
-  CompassIcon,
-  GearSixIcon,
-} from "@phosphor-icons/react";
+import { CaretDownIcon, CompassIcon } from "@phosphor-icons/react";
 import type { ScoutConfig } from "@posthog/api-client/posthog-client";
 import {
   computeFleetSummary,
   computeScoutRollups,
-  formatRunIntervalShort,
-  prettifyScoutSkillName,
-  type ScoutRollup,
-  scoutSkillSlug,
   sortConfigsForDisplay,
 } from "@posthog/core/scouts/scoutPresentation";
 import {
@@ -19,21 +10,12 @@ import {
   scoutRunsWindowLabel,
 } from "@posthog/core/scouts/scoutRunsWindow";
 import { RelativeTimestamp } from "@posthog/ui/primitives/RelativeTimestamp";
-import { skillUrl } from "@posthog/ui/utils/posthogLinks";
-import { Box, Flex, Text, Tooltip } from "@radix-ui/themes";
-import { Link } from "@tanstack/react-router";
+import { Box, Flex, Text } from "@radix-ui/themes";
 import { useMemo, useState } from "react";
 import { useScoutConfigMutations } from "../hooks/useScoutConfigMutations";
 import { useScoutConfigs } from "../hooks/useScoutConfigs";
 import { useScoutRuns } from "../hooks/useScoutRuns";
-import {
-  DryRunBadge,
-  deriveScoutRowState,
-  ScoutOriginBadge,
-  ScoutStatusDot,
-} from "./ScoutBadges";
-import { ScoutConfigForm, ScoutEnabledSwitch } from "./ScoutConfigControls";
-import { ScoutRunBoxes } from "./ScoutRunBoxes";
+import { ScoutRowCard } from "./ScoutRowCard";
 
 /**
  * Expandable scout fleet manager for the agents config page. Collapsed it is
@@ -149,7 +131,7 @@ function ScoutsFleetList({ configs }: { configs: ScoutConfig[] }) {
       <div className="max-h-[710px] overflow-y-auto">
         <Flex direction="column" gap="2">
           {visibleConfigs.map((config) => (
-            <ScoutRow
+            <ScoutRowCard
               key={config.id}
               config={config}
               rollup={rollups.get(config.skill_name)}
@@ -165,98 +147,6 @@ function ScoutsFleetList({ configs }: { configs: ScoutConfig[] }) {
         <span className="font-mono text-[11px]">signals-scout-*</span> skills in
         your PostHog project.
       </Text>
-    </Flex>
-  );
-}
-
-function ScoutRow({
-  config,
-  rollup,
-  onUpdate,
-}: {
-  config: ScoutConfig;
-  rollup: ScoutRollup | undefined;
-  onUpdate: (
-    configId: string,
-    updates: Parameters<
-      ReturnType<typeof useScoutConfigMutations>["updateConfig"]
-    >[1],
-  ) => void;
-}) {
-  const now = new Date();
-  const state = deriveScoutRowState(config, rollup, now);
-  const [settingsOpen, setSettingsOpen] = useState(false);
-  const cloudSkillUrl = skillUrl(config.skill_name);
-
-  return (
-    <Flex
-      direction="column"
-      className={`group rounded-(--radius-3) border border-border bg-(--color-panel-solid) px-4 py-3 transition duration-150 hover:border-(--gray-6) hover:bg-(--gray-2) ${
-        config.enabled ? "" : "opacity-65"
-      }`}
-    >
-      <Flex align="center" gap="4">
-        <Flex align="center" gap="2" className="min-w-0 flex-1">
-          <Link
-            to="/code/agents/scouts/$skillName"
-            params={{ skillName: scoutSkillSlug(config.skill_name) }}
-            className="flex min-w-0 items-center gap-2 no-underline"
-          >
-            <ScoutStatusDot state={state} />
-            <Text className="truncate font-medium text-[13px] text-gray-12">
-              {prettifyScoutSkillName(config.skill_name)}
-            </Text>
-          </Link>
-          {cloudSkillUrl ? (
-            <Tooltip content="View skill in PostHog">
-              <a
-                href={cloudSkillUrl}
-                target="_blank"
-                rel="noreferrer"
-                aria-label={`${config.skill_name} skill in PostHog`}
-                className="text-gray-9 opacity-0 transition-opacity hover:text-accent-11 focus-visible:opacity-100 group-hover:opacity-100"
-              >
-                <ArrowSquareOutIcon size={12} />
-              </a>
-            </Tooltip>
-          ) : null}
-          <ScoutOriginBadge skillName={config.skill_name} />
-          <DryRunBadge config={config} />
-          <Text className="whitespace-nowrap text-[11px] text-gray-10">
-            {formatRunIntervalShort(config.run_interval_minutes)}
-          </Text>
-          {rollup && rollup.emittedCount > 0 ? (
-            <Text className="whitespace-nowrap text-[11px] text-gray-10">
-              · {rollup.emittedCount} signal
-              {rollup.emittedCount === 1 ? "" : "s"} emitted
-            </Text>
-          ) : null}
-        </Flex>
-        <ScoutRunBoxes runs={rollup?.runs ?? []} />
-        <Flex align="center" gap="3" className="shrink-0">
-          <ScoutEnabledSwitch config={config} onUpdate={onUpdate} />
-          <Tooltip content="Scout settings">
-            <button
-              type="button"
-              onClick={() => setSettingsOpen((value) => !value)}
-              aria-expanded={settingsOpen}
-              aria-label={`${config.skill_name} settings`}
-              className={`flex h-6 w-6 items-center justify-center rounded transition-colors ${
-                settingsOpen
-                  ? "bg-(--gray-4) text-gray-12"
-                  : "text-gray-10 hover:bg-(--gray-3) hover:text-gray-12"
-              }`}
-            >
-              <GearSixIcon size={14} />
-            </button>
-          </Tooltip>
-        </Flex>
-      </Flex>
-      {settingsOpen ? (
-        <Box className="mt-3 border-(--gray-4) border-t pt-3">
-          <ScoutConfigForm config={config} onUpdate={onUpdate} />
-        </Box>
-      ) : null}
     </Flex>
   );
 }
